requirements.txt
langchain==0.2.16
langchain-community==0.2.16
langchain-google-genai==1.0.10
langchain-huggingface==0.0.3
faiss-cpu==1.8.0
PyPDF2==3.0.1
fastapi==0.111.0
uvicorn==0.30.1
nltk==3.8.1
scikit-learn==1.5.0
google-generativeai==0.7.2
python-dotenv==1.0.1